Move rule graphql-syntax to own file
